+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## VERSION 0.2.0 - 2024-xx-xx
 
 - Updated edges to be more usable, including class names (to list the correct list nodes)
+- Unique validation for nodes
 
 ## VERSION 0.1.0 - 2024-05-21